requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.11
django-cors-headers==4.3.1
django-dotenv==1.4.2
filelock==3.13.3
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.6
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
requests-file==2.0.0
sqlparse==0.4.4
tldextract==5.1.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
zope.event==5.0
zope.interface==6.2
